requirements.txt
# DAX 40 AI-Analyse – Python Abhängigkeiten
# Installation: pip install -r requirements.txt

# PDF-Extraktion
pymupdf==1.24.0          # fitz – schnell und zuverlässig für deutsche PDFs
pdfplumber==0.11.0       # Alternative, gut für Tabellen

# Datei-Handling & Konfiguration
pyyaml==6.0.1
python-dotenv==1.0.0

# Datenverarbeitung
pandas==2.2.0
numpy==1.26.4
pyarrow==15.0.0          # Für .parquet-Format

# Text-Analyse
nltk==3.8.1
spacy==3.7.4
# Sprachmodelle: python -m spacy download de_core_news_lg
# Sprachmodelle: python -m spacy download en_core_web_sm

# N-Gram & Statistik
scikit-learn==1.4.0
scipy==1.12.0

# Visualisierung
matplotlib==3.8.3
seaborn==0.13.2
plotly==5.18.0           # Interaktive Charts für Exploration

# Web-Scraping (für Downloader)
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.18.1         # Für JS-heavy IR-Seiten
webdriver-manager==4.0.1

# Jupyter
jupyterlab==4.1.0
ipywidgets==8.1.2

# Fortgeschrittene Analyse (optional, Phase 2)
# transformers==4.38.0   # HuggingFace für Sentiment
# bertopic==0.16.2       # Topic Modeling
